+++ Firestore/Source/Public/FirebaseFirestore/FirebaseFirestore.h
@@ -30,6 +30,7 @@
 #import "FIRGeoPoint.h"
 #import "FIRListenerRegistration.h"
 #import "FIRLoadBundleTask.h"
+#import "FIRLocalCacheSettings.h"
 #import "FIRQuery.h"
 #import "FIRQuerySnapshot.h"
 #import "FIRSnapshotMetadata.h"