+++ Firestore/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 10.9.0
 - [feature] Add new cache config API to customize SDK cache settings.
 - [feature] Add LRU garbage collector as an option to memory cache.